requirements.txt
scikit_image==0.16.2
Pillow==7.1.2
numpy==1.19.5
torchvision==0.9.1+cu111
torch==1.8.1+cu111
